kernel: ksud: provide ksu_handle_compat_execve_ksud
Since `kernel/selinux: fix pointer mismatch with 32-bit ksud on 64-bit kernels`
fixed 32-on-64 setups, it is now possible to run KernelSU properly on those.

But an issue arises around its exec hook! sys_execve passes native pointers.
Good -- thats what we want. But then on 32-on-64 setups, its not a good idea.
ksud will still be ran on sys_execve!

Here we provide a helper function for it, ksu_handle_compat_execve_ksud.
usage is the same as ksu_handle_execve_ksud, the only difference is that
for argv, it uses argv.ptr.compat instead of argv.ptr.native.

Yes this is possible to commonize and we do that on next commit.

Signed-off-by: backslashxx <118538522+backslashxx@users.noreply.github.com>